Moved helper functions into utils folder
