+++ src/popup.tsx
@@ -4,7 +4,7 @@ import { CountryDropdown } from "./components/CountryDropdown";
 import { usePopup } from "./usePopup";
 import { SelectedCountries } from "./components/selected-countries/SelectedCountries";
 import { CalculationHistory } from "./components/calculation-history/CalculationHistory";
-import { formatBytesString } from "./helpers/formatBytesString";
+import { formatBytesString } from "./utils/helpers/formatBytesString";
 
 export const Popup = () => {
     const {
