+++ src/usePopup.ts
@@ -1,9 +1,9 @@
 import { useEffect, useState } from "react";
 import { CountryName } from "./data/constants/CountryEmissions";
-import { calculateAverageSpecificEmissionsHelper } from "./helpers/calculateAverageSpecificEmissions";
-import { calculateCarbon } from "./helpers/calculateCarbon";
+import { calculateAverageSpecificEmissionsHelper } from "./utils/helpers/calculateAverageSpecificEmissions";
+import { calculateCarbon } from "./utils/helpers/calculateCarbon";
 import { ISelectedCountriesRepository } from "./data/selected-countries/ISelectedCountriesRepository";
-import { useMountEffect } from "./helpers/useOnceAfterFirstMount";
+import { useMountEffect } from "./utils/helpers/useOnceAfterFirstMount";
 import {
     CalculationData,
     ICalculationsRepository,
