+++ src/utils/helpers/__tests__/areMapsDeepEqual.test.ts
@@ -1,4 +1,4 @@
-import { CountryName } from "../../data/constants/CountryEmissions";
+import { CountryName } from "../../../data/constants/CountryEmissions";
 import { areMapsDeepEqual } from "../areMapsDeepEqual";
 
 const mockMap1: Map<CountryName, number> = new Map([
