+++ src/utils/helpers/__tests__/calculateCarbon.test.ts
@@ -1,4 +1,4 @@
-import { CountryName } from "../../data/constants/CountryEmissions";
+import { CountryName } from "../../../data/constants/CountryEmissions";
 import { calculateCarbon } from "../calculateCarbon";
 
 const mockCountry: Map<CountryName, number> = new Map([["Australia", 0.5]]);
